Move web components into separate gradle module
