+++ app/src/main/java/com/buddybuild/di/MainComponent.java
@@ -1,7 +1,6 @@
 package com.buddybuild.di;
 
-
-import com.buddybuild.ui.AppsActivity;
+import com.buddybuild.rest.RestModule;
 import com.buddybuild.ui.AppsFragment;
 
 import javax.inject.Singleton;
@@ -12,9 +11,8 @@ import dagger.Component;
  * Dagger component for main package
  */
 @Singleton
-@Component(modules = {MainModule.class})
+@Component(modules = {MainModule.class, RestModule.class})
 public interface MainComponent {
-    void inject(AppsActivity activity);
 
     void inject(AppsFragment appsFragment);
 }
