+++ rest/src/main/java/com/buddybuild/rest/AppResponse.java
