+++ rest/src/main/java/com/buddybuild/rest/WebService.java
