+++ rest/src/test/java/com/buddybuild/rest/AppResponseTest.java
@@ -1,0 +1,11 @@
+package com.buddybuild.rest;
+
+import org.junit.Test;
+
+
+public class AppResponseTest {
+    @Test
+    public void appResponseCanMapToApp() throws Exception {
+        // TODO
+    }
+}
